refactor eval
